+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## [1.0.2](https://github.com/US-EPA-CAMD/easey-design-system/compare/v1.0.1...v1.0.2) (2021-11-06)
+
+
+### Bug Fixes
+
+* removing useState from components ([cc5cfc1](https://github.com/US-EPA-CAMD/easey-design-system/commit/cc5cfc1dc18b49f0e51a4f0c77111514cb3bf423))
+
+## [1.0.1](https://github.com/US-EPA-CAMD/easey-design-system/compare/v1.0.0...v1.0.1) (2021-11-06)
+
+
+### Bug Fixes
+
+* working through some issues ([40a48a1](https://github.com/US-EPA-CAMD/easey-design-system/commit/40a48a1513a4f48f5472240bd8bd3b289ac3935c))
+
+# 1.0.0 (2021-11-05)
+
+
+### Bug Fixes
+
+* added auth token to workflow for publish-storybook ([a803e39](https://github.com/US-EPA-CAMD/easey-design-system/commit/a803e393b74b5c3a31a082f7571b9f460fc8ca66))
+* initial release of easey-design-system ([a0febdb](https://github.com/US-EPA-CAMD/easey-design-system/commit/a0febdbbda1c3e388f678f1cd5ad022b5a80a7d0))
+* removed publish-storybook ([4bd557c](https://github.com/US-EPA-CAMD/easey-design-system/commit/4bd557cb7c89e976eda19493e4dcffeeed3b9ab0))
+
+
+### Features
+
+* 1st version of easey-design-system ([5554bb9](https://github.com/US-EPA-CAMD/easey-design-system/commit/5554bb994b10fb708f87090410d664cd4cf77e8a))
+* finalized easey-design-system package ([2daa461](https://github.com/US-EPA-CAMD/easey-design-system/commit/2daa46173baf787257ae99610edf8b427a5a1d11))
+* gh-pages added ([0a9bd5b](https://github.com/US-EPA-CAMD/easey-design-system/commit/0a9bd5b7a5207f57105a1dce22373b82f4ba996d))
+
 # 1.0.0 (2021-11-05)